+++ lib/karafka/pro/connection/manager.rb
@@ -23,44 +23,13 @@ module Karafka
         include Core::Helpers::Time
 
         # How long should we wait after a rebalance before doing anything on a consumer group
-        def initialize(change_delay: 1 * 1_000)
+        # @param change_delay [Integer] what should be the delay before applying any changes after
+        #   last change on a consumer group. This allows us for stabilization so we do not cause
+        #   many rebalances too often. By default it is one minute.
+        def initialize(change_delay: 60 * 1_000)
           super()
           @change_delay = change_delay
           @changes = Hash.new { |h, k| h[k] = monotonic_now }
-        end
-
-        # Registers listeners and starts the scaling procedures
-        #
-        # @param listeners [Connection::ListenersBatch]
-        def register(listeners)
-          @listeners = listeners
-
-          in_families do |subscription_group, multiplexing, sg_listeners|
-            notice(subscription_group.consumer_group)
-
-            if multiplexing.active? && multiplexing.dynamic?
-              # Start absolute minimum that is expected in this multiplexed group for now and leave
-              # the upscaling for later. This will minimize number of connections in total in
-              # all consumer groups effectively allowing for faster start
-              sg_listeners[0...multiplexing.min].each(&:start)
-            else
-              sg_listeners.each(&:start)
-            end
-          end
-        end
-
-        def in_families
-          grouped = @listeners.group_by { |listener| listener.subscription_group.name }
-
-          grouped.each_value do |listeners|
-            listener = listeners.first
-
-            yield(
-              listener.subscription_group,
-              listener.subscription_group.multiplexing,
-              listeners
-            )
-          end
         end
 
         # Marks most recent time on a given consumer group. This is used to back-off any
@@ -68,11 +37,6 @@ module Karafka
         # @param consumer_group_id [String] consumer group id
         def notice(consumer_group_id)
           @changes[consumer_group_id] = monotonic_now
-        end
-
-        def scale
-          # Do nothing if we just started to stop the process
-          return if Karafka::App.done?
         end
       end
     end