+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Karafka framework changelog
 
-## 2.2.15 (Unreleased)
+## 2.3.0 (Unreleased)
 - **[Feature]** Provide ability to multiplex subscription groups (Pro)
 - **[Feature]** Provide `Karafka::Admin::Acl` for Kafka ACL management via the Admin APIs.
 - **[Feature]** Periodic Jobs (Pro)
 - **[Feature]** Offset Metadata storage (Pro)
+- **[Feature]** Provide low-level listeners management API for dynamic resources scaling (Pro)
+- [Enhancement] Improve shutdown process.
 - [Enhancement] Introduce `non_blocking` routing API that aliases LRJ to indicate a different use-case for LRJ flow approach.
 - [Enhancement] Allow to reset offset when seeking backwards by using the `reset_offset` keyword attribute set to `true`.
 - [Enhancement] Alias producer operations in consumer to skip `#producer` reference.